+++ db/migrate/20160829135248_add_ignore_auto_lock_to_designs.rb
@@ -1,0 +1,5 @@
+class AddIgnoreAutoLockToDesigns < ActiveRecord::Migration[5.0]
+  def change
+    add_column :designs, :ignore_auto_lock, :boolean, null: false, default: false
+  end
+end