+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - **Data Quality Checks**
   - Project editors can now define data quality checks that are used to identify
     sheets that have inconsistent data
+- **Design Changes**
+  - Designs can be flagged to ignore project-wide auto-locking
 - **Sheet Changes**
   - Sheet saving errors are now written to a hidden field on the form to allow
     debugging of branching logic mistakes